snowcamp-docling/requirements.txt
accelerate>=1.0.0
docling>=2.0.0
docling-core>=2.0.0
docling-parse>=4.0.0
ipykernel>=6.30.0
ipywidgets>=8.0.0
jupyter>=1.0.0
langchain>=0.1.0
langchain-chroma>=0.1.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-huggingface>=0.3.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.3.0
openai>=1.0.0
pydantic>=2.0.0
PyPDF2>=3.0.0
pycryptodome>=3.20.0
python-dotenv>=1.0.0
pypdfium2>=4.30.0
sentence-transformers>=5.0.0
tiktoken>=0.11.0
torch>=2.0.0
transformers>=4.0.0
